FIT-37: Update JwtConfig to use long instead of int

+++ src/main/java/com/fittracker/fittracker/config/Configuration.java
@@ -9,6 +9,6 @@ public class Configuration {
     @Bean
     public JwtConfig jwtConfig(@Value("${security.jwt.secret}") String secret,
         @Value("${security.jwt.tokenExpirationPeriodMinutes}") String tokenExpirationPeriodMinutes) {
-        return new JwtConfig(secret, Integer.parseInt(tokenExpirationPeriodMinutes));
+        return new JwtConfig(secret, Long.parseLong(tokenExpirationPeriodMinutes));
     }
 }
